Fixed issue with words broken up by HTML tags.
To address this we now pass a custom 'replace' function to
findAndReplaceDomText, which allows us to generate custom HTML for the
highlight elements. The word to be replaced is added to these elements
as data, and the suggestions will be based off of this rather than the
text content of the span.

Also fixed an issue where the highlighting was not properly removed upon
ignoring words or cancelling the spellchecker.